Fix PPC compile failures: pointer/array, void* arithmetic, null derefs

+++ tests/end_to_end/large-struct-offset/agbcc-o2-out.c
@@ -1,5 +1,5 @@
 extern s32 glob;
 
 void test(void *arg0) {
-    glob = arg0->unk12348;
+    glob = (s32) arg0->unk12348;
 }
